feat: refactor page rendering by introducing PageRenderManager and updating IPage interface
